refactor: Move `Identifier` to shared API model
Components should not depend on the legacy API model but the shared API
model instead.
