+++ api/v1/model/src/commonMain/kotlin/Issue.kt
@@ -24,6 +24,7 @@ import kotlin.time.Instant
 import kotlinx.serialization.Serializable
 
 import org.eclipse.apoapsis.ortserver.shared.apimodel.AppliedIssueResolution
+import org.eclipse.apoapsis.ortserver.shared.apimodel.Identifier
 import org.eclipse.apoapsis.ortserver.shared.apimodel.IssueResolution
 
 /**
