+++ api/v1/model/src/commonMain/kotlin/Package.kt
@@ -21,6 +21,8 @@ package org.eclipse.apoapsis.ortserver.api.v1.model
 
 import kotlinx.serialization.Serializable
 
+import org.eclipse.apoapsis.ortserver.shared.apimodel.Identifier
+
 @Serializable
 data class Package(
     val identifier: Identifier,
