+++ api/v1/model/src/commonMain/kotlin/RuleViolation.kt
@@ -22,6 +22,7 @@ package org.eclipse.apoapsis.ortserver.api.v1.model
 import kotlinx.serialization.Serializable
 
 import org.eclipse.apoapsis.ortserver.shared.apimodel.AppliedRuleViolationResolution
+import org.eclipse.apoapsis.ortserver.shared.apimodel.Identifier
 import org.eclipse.apoapsis.ortserver.shared.apimodel.RuleViolationResolution
 
 @Serializable
