+++ components/dependency-graph/api-model/src/commonMain/kotlin/DependencyGraph.kt
@@ -21,7 +21,7 @@ package org.eclipse.apoapsis.ortserver.components.dependencygraph
 
 import kotlinx.serialization.Serializable
 
-import org.eclipse.apoapsis.ortserver.api.v1.model.Identifier
+import org.eclipse.apoapsis.ortserver.shared.apimodel.Identifier
 
 @Serializable
 data class DependencyGraphs(
