+++ components/search/api-model/src/commonMain/kotlin/RunWithPackage.kt
@@ -23,7 +23,7 @@ import kotlin.time.Instant
 
 import kotlinx.serialization.Serializable
 
-import org.eclipse.apoapsis.ortserver.api.v1.model.Identifier
+import org.eclipse.apoapsis.ortserver.shared.apimodel.Identifier
 
 @Serializable
 data class RunWithPackage(
