+++ services/ort-run/src/main/kotlin/VulnerabilityService.kt
@@ -24,7 +24,6 @@ import com.github.michaelbull.result.getOr
 import java.util.SortedSet
 
 import org.eclipse.apoapsis.ortserver.api.v1.model.AdvisorDetails
-import org.eclipse.apoapsis.ortserver.api.v1.model.Identifier
 import org.eclipse.apoapsis.ortserver.api.v1.model.Vulnerability
 import org.eclipse.apoapsis.ortserver.api.v1.model.VulnerabilityRating
 import org.eclipse.apoapsis.ortserver.api.v1.model.VulnerabilityReference
@@ -58,6 +57,7 @@ import org.eclipse.apoapsis.ortserver.services.ResourceNotFoundException
 import org.eclipse.apoapsis.ortserver.services.utils.toSortOrder
 import org.eclipse.apoapsis.ortserver.shared.apimappings.mapToApi
 import org.eclipse.apoapsis.ortserver.shared.apimodel.AppliedVulnerabilityResolution
+import org.eclipse.apoapsis.ortserver.shared.apimodel.Identifier
 import org.eclipse.apoapsis.ortserver.shared.apimodel.ResolutionSource
 import org.eclipse.apoapsis.ortserver.shared.apimodel.VulnerabilityResolution
